Grandfather-at-birth: split src/runs/stream/bridge.test.ts under 500 lines (warren-889a)
