+++ .mulch/expertise/build.jsonl
@@ -109,3 +109,4 @@
 {"type":"failure","classification":"tactical","recorded_at":"2026-06-20T22:28:08.996Z","evidence":{"commit":"ec65e1823ef92ed5249ca8a1b4ad77434cdf1b1d"},"description":"release.md patch flow: a prior run dropped the '## [0.9.3]' CHANGELOG header, merging that release's content into [Unreleased], so the next bump would re-document already-released items.","resolution":"Before moving [Unreleased] into the new dated section, diff CHANGELOG vs the last tag (git diff vX..HEAD -- CHANGELOG.md) to identify the genuinely-new entries; restore any dropped prior-version header and move only the new entries. After bumping VERSION, re-run 'bun run gen:openapi' since docs/openapi.yaml info.version follows src/index.ts VERSION (gen:openapi:check fails otherwise).","id":"mx-b54298"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-06-20T23:38:53.322Z","evidence":{"commit":"dd83e590b98cac2ce76f2e5ac6d468e0dae3840c"},"content":"check:size ratchet: src/server/bridge-reconnect.ts was exactly AT the 500-line default threshold and src/runs/reap/run.ts at its frozen 527 budget, so any addition trips check:size. For warren-9f06 the fix was extracting the two self-contained per-project config resolvers (resolveProjectPreviewConfig/resolveProjectPrTemplate) out of bridge-reconnect.ts into a new sibling src/server/bridge-reconnect-config.ts (type-only circular import of RunWithReconnectInput is fine), matching that file's own 'extracted to keep under the file-size ratchet' ethos, rather than raising the budget. For reap/run.ts shave lines by collapsing multi-line log objects to single lines under 100c and hoisting err->reason locals.","id":"mx-70175b"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-06-20T23:59:37.977Z","evidence":{"commit":"ffc146aa4b987e551fccaf8f3e9e752c3b95b906"},"dir_anchors":["src/server/main"],"content":"Central pino redaction (warren-b2dd/pl-f700 step 6): src/server/main/redact.ts exports LOG_REDACT_OPTIONS (paths + censor '[Redacted]') covering token-shaped fields (token/githubToken/GITHUB_TOKEN/BURROW_API_TOKEN/authorization/bearer etc.) bare, one-level-deep (*.field), and headers.authorization. Applied via createWarrenLogger(env) in src/server/main/logging.ts (the single pino factory, replaces inline pino() in index.ts) and re-imported in src/supervisor/main.ts. pino redact paths are NOT deep-recursive so the wildcard tier is enumerated explicitly.","id":"mx-98f2ad"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-06-22T03:53:50.529Z","evidence":{"commit":"48f6add3e5d8026beb33eddc82ba480969fbdaf1"},"content":"warren-889a / pl-c34d step 5: src/runs/stream/bridge.test.ts (521 lines, grandfathered-at-birth in be18ba73/warren-df71) split below the 500-line check:size budget. The in-stream terminal-detection + conversation keep-alive describe blocks moved to sibling src/runs/stream/bridge.terminal.test.ts; original keeps the 'event flow' describe. Budget entry removed from scripts/file-size-budgets.json (ratchet only goes down). Sibling file repeats the seedBridgeRun beforeEach/afterEach boilerplate; unused BurrowNotFoundError import dropped from the trimmed original.","id":"mx-a259b0"}
